Refactor the Gameboard system
